+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/RandomDotOrgAnonymousClientLiveTest.java
@@ -1,0 +1,7 @@
+package io.github.pr0methean.betterrandom.seed;
+
+public class RandomDotOrgAnonymousClientLiveTest extends WebSeedClientLiveTest<RandomDotOrgAnonymousClient> {
+  protected RandomDotOrgAnonymousClientLiveTest() {
+    super(RandomDotOrgAnonymousClient.WITHOUT_DELAYED_RETRY);
+  }
+}