+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/SeedGeneratorTest.java
@@ -2,15 +2,28 @@ package io.github.pr0methean.betterrandom.seed;
 
 import static org.testng.Assert.assertEquals;
 
+import javax.annotation.Nullable;
+import org.powermock.modules.testng.PowerMockTestCase;
 import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-public abstract class SeedGeneratorTest<T extends SeedGenerator> {
+public abstract class SeedGeneratorTest<T extends SeedGenerator> extends PowerMockTestCase {
 
   protected T seedGenerator;
 
+  @Nullable protected final T defaultSeedGenerator;
+
   protected SeedGeneratorTest(final T seedGenerator) {
-    this.seedGenerator = seedGenerator;
+    this.defaultSeedGenerator = seedGenerator;
+  }
+
+  protected T getSeedGenerator() {
+    return defaultSeedGenerator;
+  }
+
+  @BeforeMethod public void setUp() {
+    seedGenerator = getSeedGenerator();
   }
 
   @Test public void testToString() {