+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/WebSeedClientLiveTest.java
@@ -6,15 +6,16 @@ import java.io.IOException;
 import java.net.Proxy;
 import java.net.URL;
 import org.testng.SkipException;
+import org.testng.annotations.AfterMethod;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeSuite;
 import org.testng.annotations.Test;
 
-public abstract class WebJsonSeedGeneratorLiveTest<T extends WebJsonSeedGenerator>
+public abstract class WebSeedClientLiveTest<T extends WebSeedClient>
     extends SeedGeneratorTest<T> {
   protected final Proxy proxy = SeedTestUtils.createProxy();
 
-  protected WebJsonSeedGeneratorLiveTest(T seedGenerator) {
+  protected WebSeedClientLiveTest(T seedGenerator) {
     super(seedGenerator);
   }
 
@@ -27,6 +28,10 @@ public abstract class WebJsonSeedGeneratorLiveTest<T extends WebJsonSeedGenerato
     }
   }
 
+  @Test public void testGenerator() {
+    SeedTestUtils.testGenerator(seedGenerator, true);
+  }
+
   @Test public void testSetProxyReal() {
     try {
       new URL("https://google.com").openConnection(proxy).getContent();
@@ -35,17 +40,19 @@ public abstract class WebJsonSeedGeneratorLiveTest<T extends WebJsonSeedGenerato
     }
     seedGenerator.setProxy(proxy);
     try {
-      SeedTestUtils.testGenerator(seedGenerator, true);
+      testGenerator();
     } finally {
       seedGenerator.setProxy(null);
     }
   }
 
-  @BeforeSuite public void setUpSuite() {
+  @Override protected T getSeedGenerator() {
+    T seedGenerator = super.getSeedGenerator();
     seedGenerator.setSslSocketFactory(createSocketFactory()); // run all tests with POODLE protection
+    return seedGenerator;
   }
 
-  @AfterSuite public void tearDownSuite() {
+  @AfterMethod public void tearDownSuite() {
     seedGenerator.setSslSocketFactory(null);
   }
 }